requirements.txt

pyresparser
streamlit
pandas==1.3.5
plotly
pymysql
streamlit-tags
Pillow
youtube-dl
nltk
pdfminer.six
spacy==2.3.9
numpy==1.19.5
python-dotenv
PyMuPDF
google-generativeai
requests
beautifulsoup4
